[kernel] bcma: update to version from wireless-testing tag master-2012-07-16
I did not add the last patch, because it add some stubs for the flash drivers and our patches have to be changed for that, if the flash support in bcma is able to do something I will integrate it.

This also contains 3 additional patches fixing some bugs in bcma.

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@32880 3c298f89-4303-0410-b956-a3cf2f4a3e73