[AR3k Bluetooth] Add support for QCA AR3k Bluetooth module
Added a new menu choice to Kernel Modules->Other modules to
build the ath3k bluetooth driver and copy the necessary dfu
and ramps firmware files to enable the device.

Tested with the current TOT Bluez stack to verify that the
module is loaded correctly and works to a basic level.

Change-Id: Iccf1113074e0cdac6b03c6f2bcfab3e261581e2c
Signed-off-by: Michael Larson <mlarson@codeaurora.org>
Reviewed-by: Mike Baker <bmichael@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>