sch_hfsc: change_class can remove service curves
The existing implementation of hfsc_change_class permits the addition of
service curves by passing non-zero SCs.  However, after service curves
are added, they cannot be removed.  Passing a NULL for a service curve
does not change it, and passing m1/m2 == 0 is treated just like a NULL.

This patch treats the m1/m2 == 0 case as a request for "zeroing" the
curve.  If the curve exists, it is removed.  The NULL case is left in
it's current state so that is is compatible with the behavior of tc
today.

Change-Id: Id2b03066865f78e986f51980017ca3d513603dfa
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-by: Matthew Mcclintock <mmcclint@codeaurora.org>