[feeds] add a generic streamboost customer feed
This is for development and it contains generic components for a
typical streamboost customer. It would be replaced with a customer
specific feed but for final images but that will be done via
build scripts which produce the final output and not via repo
and/or the manifest file. This default feed should be enough
for the average developer to complete their work.

Change-Id: Ifb7b75ab9f64ff6a50ccd32326d7843aaa0576ae
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>