[qca-nss-drv] Do not git-fetch version in local-dev mode
We don't have to fetch the version from the git repo when
working in local development mode.
This results in a failure if the git repo is unavailable.

After this change, the package will now git-fetch the version
only if we detect that the local dev tree is not present

Change-Id: Ic711224e53ccf7f7b46dee6b22b396fb03ddc112
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-by: Mike Baker <bmichael@codeaurora.org>
Reviewed-by: Matthew Mcclintock <mmcclint@codeaurora.org>