[qca-nss-gmac] Do not git-fetch version in local-dev mode
We don't have to fetch the version from the git repo when
working in local development mode.
This results in a failure if the git repo is unavailable.

After this change, the package will now git-fetch the version
only if we detect that the local dev tree is not present

Change-Id: I56ca0c798531a45ff177f217884b0855846da5d9
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>