[qca-nss-*] Clean up makefile
We can't use "?=" with a shell command since that will evaluate the
shell command each reference to the variable. Switched to a ":="
wrapped with an ifeq. Additionally check the DUMP variable to see
if we're generating a make menuconfig list, if so skip the shell
command since the version doesn't matter in the menu.

Change-Id: Ic0353170763f4e3873b59ea0d543eef031f449d4
Signed-off-by: Mike Baker <bmichael@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>