[netdev] generic skb_recycling
Defines a simplistic RX-to-RX skb recycling scheme for bridging and
routing workloads. This version is designed to work with default RX
buffers up to 2KB in size. We can certainly do pools of power-of-2
sizes, should the need arise.

The downside of this approach is for skb allocations that are
significantly smaller than 2K; these become 2K allocations. However,
this is relatively rare for the two functions we have hooked.

We have used a per-CPU list for the SKBs, and we disable preemption
while accessing the list.

Overall, the performance gain on the wired-to-wireless bridging
workload was around 16%.

Change-Id: Id6cb909422f1c229cfd34aca55514cabd27349df
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>
Reviewed-by: Benjamin Menchaca <bmenchac@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>