[mac80211] update db.txt to upstream bb99560
Part of a full fix for EV 129103

In EV 129103 the PVT reports they a way to test
regulatory. We need a short term solution and a long
term solution. This is part of a fix for the bug report.

The short term solution involves the following. This commit
only deals with 1). Below are details of the overall
plan and explanation of how OpenWrt / QSDK deals with
regulatory for documentation purposes.

  1) Update the QSDK db.txt to match upstream
  2) Update the QSDK db.txt to match VHT80 updates,
     and send the VHT80 changes upstream
  3) Enable PVT to test regulatory changes with a
     private patch
  4) Fix any issues reported by PVT on db.txt and send
     fixes upstream

The Long term solution involves:

  a) QSDK rebases itself to latest OpenWrt which uses
     the new backports project, instead of the compat-wireless
     package. QSDK team provides PVT specific images to PVT
     with CONFIG_CFG80211_CERTIFICATION_ONUS enabled and
     other options that depend on CONFIG_CFG80211_CERTIFICATION_ONUS
     depending on agreement between the QSDK team and PVT.
     QSDK team should be careful to not enalbe options that
     depend on CONFIG_CFG80211_CERTIFICATION_ONUS on production
     images without verying with the regulatory team.

  b) QCA QSDK team works with regulatory team, Michael Green,
     to do automated scripted updates of db.txt from
     the regulatry domain regDomain.xlsx file converted to
     Open Document format, regDomain.odf. The changes should
     be sent upstream for review / integration. There are other
     teams at QCA that depend on this work and as such this work
     should be coordinated within QCA. Example teams that also
     need these updates are the WCN team.

  c) A patch is developed to enable PVT to enable testing
     regulatory for the ath module which will depend on
     CONFIG_CFG80211_CERTIFICATION_ONUS. This patch should
     be submitted upstream upon review with QCA regulatory
     team. The CONFIG_CFG80211_CERTIFICATION_ONUS option
     was already vetted by QCA regulatory for integration
     upstream in order to support cellular base station hint
     support (see Linux upstream commit 57b5ce072e).

The way OpenWrt deals with regulatory currently is it
provides an update to the 3.3 base kernel with an updated
cfg80211 as part of the "mac80211" package. This mac80211
package is essentially the compat-wireles package, now
renamed to "backports". The new "mac80211" package upgrades
the cfg80211 module and as such updates the base kernel's
802.11 regulatory component which is dealt with in cfg80211.

cfg80211 allows for regulatory updates via two mechanisms:

  * udev hints triggering CRDA to send updates from wireless-regdb
  * in-kernel updates when CONFIG_CFG80211_INTERNAL_REGDB is enabled

By default OpenWrt chooses to use CONFIG_CFG80211_INTERNAL_REGDB
and ships the wireless-regdb db.txt as part of the "mac80211"
package and embeds it in package/mac80211/files/regdb.txt. Upon
build time OpenWrt copies over the regdb.txt to the compat-wireless
net/wireless/db.txt. This embeds the regulatory databse as part
of the kernel. OpenWrt and QSDK by default disable the CRDA
package however therefore requiring a full image build if a
regulatory update is provided. For the short term we'll be
updating then package/mac80211/files/regdb.txt and providing
PVT team with a private patch to enable testing of regulatory.

Change-Id: I4713326ec6d59d6156803e266123a525ec372ea7
Signed-off-by: Luis R. Rodriguez <rodrigue@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>