[ipq806x] Add msm-3.4 config opts to generic/config-3.4
Some new options have been added to the kernel due to either
the 3.4 upgrade or to the fact that it's got the Android
modifications.

Either way, these options needs to have a default value in
config-3.4 so that when the package get selected these get
enabled.

The added options here are related to Netfilter and PPP:

Change-Id: Ib93ce1b7936ac68e95f251f4af8d1e4ee45e3c45
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>