[iptables] Apply patches even in ext. kernel tree case
This commit reverts the following previous change:

40faa2b2ed5f66f91d8ee98855e4bb8079671c01
[package] do not apply iptables patches when building with CONFIG_EXTERNAL_KERNEL_TREE

The rule is wrong. By default external kernels should be patched to
support the features available in OpenWrt. The same problem will
happen in every single package and testing EXTERNAL_KERNEL_TREE
is not scalable.

Having the corresponding iptables patches in the kernel should be
a requirement to enable the iptables package. Else, it should be
up to the ext. kernel tree maintainer to take care of this problem.

Change-Id: I9a66c82907bf156f4896d0a1b200452321143d00
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>