[include] kernel: add LINUX_SRC_DIR variable
In case where an external kernel is used and O= option is used,
we may end-up having the sources in one directory and the objects
in another. This is a problem for some packages that access the
kernel headers directory by pointing to the kernel directory.

To accomodate this behavior, this change adds a new variable
called LINUX_SRC_DIR that will always point to the kernel sources
whether an external kernel is used or not. If not, it will point
to the kernel in build_dir, else it will point to
CONFIG_EXTERNAL_KERNEL_TREE.

In both case, LINUX_DIR keeps pointing to the kernel in build_dir
so it doens't change anything for the existing packages.

Change-Id: Ic2d6a4a2b8dfa840965066cc1d74fc65697daedb
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>