[ag71xx] fix invalidate past end of skb
When the driver allocates SKBs, it allocates rx_buf_size.  When this
is DMA mapped, it is invalidated starting at skb->data.  Since data is
moved by rx_buf_offset, this means that rx_buf_offset data is
incorrect invalidated.  This includes the shinfo area.  This patch
increases the size of the skb allocations by rx_buf_offset.

Change-Id: I45018d9f7c56b74be53619d870328a89815f8d06
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>