[netdev] Fix skb recycle EVs 125834 125845
There were two bugs fixed here. The first is a variable name typo (
len instead of length), and len is uninitialized. The second was an
inconsistency in the allocated lengths between dev_alloc_skb, and
__netdev_alloc_skb. Now both cases add NET_SKB_PAD to the allocated
length after the comparison to RECYCLE_SIZE to account for the reserved
headroom and still provide the correct size SKB.

Original bug found and suggested fix by Xiaoyang Wang.

Change-Id: I92077eadc14cb5aff524b14fc48fdfa8407a2485
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>