[kernel] Adds MSM related USB gadget kernel module options
This change adds the opportunity to select the DWC3 USB gadget
driver from the openwrt menuconfig. It's used for IPQ806x and is
in the msm-3.4 kernel.

This change only adds the options in the menuconfig, and adds
usbgadget to the list of supported features for this platform.
However, it doesn't add these modules to the firmware configuration,
it only makes them selectable without actually selecting them.

Change-Id: I35bbd5f6e4f06082e7f2b26d2e119ee3d7588304
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>