[qca-nss-drv] Move the connection manager into a separate BuildPackage
Connection manager will not be selected for carrier router and
enterprise AP profiles. So, it needs to be selectable to be integrated
only in the profiles that we want.

This change splits out the existing qca-nss-drv package into 2
different BuildPackage, so the conn-mgr can be selected separately.
This change also adds it to the retail and open packages but not
to the carrier and enterprise that do not want the NSS accelerations.

Change-Id: Iee85850f706e4eeb7137ce0683f212125892c9b8
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>