[busybox] Do not select bridge by default if bridge-utils is enabled
Both busybox and bridge-utils can install the /usr/bin/brctl command.
When both are enabled in the config, the busybox one is used, which
is a problem as busybox enables its brctl by default.

This change is modifying the busybox default behavior so that brctl
doesn't get selected automatically if bridge-utils was already selected.
Which means that in the profile, when we will select bridge-utils
it will automatically disable brctl in the busybox.

Change-Id: Ie2adda91a31f4f549e6a68a9b9bb8af11fa67964
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>