[linux/generic] Add MSM USB gadget opts to generic kernel conf
This change adds the options CONFIG_USB_DWC3_MSM and
CONFIG_USB_DWC3_IPQ to the generic kernel configuration so it gets
automatically selected and the user doesn't get prompted when standard
gadget support is enabled but these option are not explicitely set.

It will have no impact until we enabled the gadget support in the
image.

Change-Id: If59a081cd5202b3ffbc2dbcf550bbd784beb0562
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-by: Krishnamoorthy Chakrapani <c_chakrk@codeaurora.org>
Reviewed-by: Matthew Mcclintock <mmcclint@codeaurora.org>