[ipq806x] Add ip6tables to all QSDK profiles
This module will add IPv6 support to netfilter by adding the
corresponding options to the kernel. It is required so that
the NSS hardware can accelerate IPv6 traffic.

This change also adds a config options in
target/linux/generic/config-3.4 that is present in msm-3.4 kernel
and need to be defined so the user doesn't get prompted during the
build. This option will be disabled by default.

Change-Id: Iaec5e5cecc24c7e44c79d0ce3212cc8001c1d017
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-by: Matthew Mcclintock <mmcclint@codeaurora.org>