[kernel/other] Split regmap package into 2 different BuildPackage
regmap enables both I2C and I2S. On SoC that do not have one of these
interfaces enabled, then OpenWrt will try to copy the kernel module
but it will fail as the corresponding kernel module is not generated.

So in this change we split the package so that the driver that depend
on these interfaces can select either one or the other.

Change-Id: Iaf368e3aa0d6a30b8f7214902cbdb89265b03f5d
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-by: Matthew Mcclintock <mmcclint@codeaurora.org>