[qca-nss-drv] Add one rule to accept bridged pkts when pkgs is enabled
qca-nss-drv package requires the iptables NF CT hooks to be enabled
even in bridge mode for the acceleration framework to be notified
when new rules are added.

A previous commit added these hooks when enabling the package (through
sysctl), but now all the packets going through the bridge are blocked.
This change adds a new firewall rule when this package is enabled, so
all the packets that are bridged and not routed can be ACCEPTed.

Change-Id: Id146077bd2278781454af6469ce9896e9a7cb365
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>