[mac80211] synch mac80211 package with OpenWrt r37821
This takes the mac80211 package up in sync with
the public OpenWrt r37821 [0]. This is needed to help
keep in synch ath9k for the ongoing changes that
will add MCC support to mac80211 and ath9k. The
package patches will likely need to get updated
regularly.

The OpenWrt mac80211 package now lies in package/kernel
we keep it tucked under package/mac80211 as with the
older releases.

Change-Id: I00e98c9c2571f05c7f4eb155ce5ebeab5837963a
Signed-off-by: Luis R. Rodriguez <rodrigue@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>