ath10k: enable Atheros 802.11ac driver
This enables usage of the ath10k driver on OpenWrt. Only 2.0 PCI hardware
is supported. The ath10k firmware is currently only available
through github so use that as its source for now.

This goes build tested, not run time tested.

This patch was already submitted upstream for inclusion on OpenWrt.

Change-Id: Ieb6476343bd808c7586e1f774918c7d8d6afd0f8
Cc: ath10k@lists.infradead.org
Cc: Kalle Valo <kvalo@codeaurora.org>
Cc: Kathy Giori <kgiori@codeaurora.org>
Cc: Mathieu Olivari <mathieu@codeaurora.org>
Cc: Jerome Proffit <jproffit@codeaurora.org>
Cc: Sujith Manoharan <c_manoha@codeaurora.org>
Cc: Matthew McClintock <mmcclint@codeaurora.org>
Signed-off-by: Luis R. Rodriguez <rodrigue@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>