add V=e option to show start and stop for tasks
Sometimes when running multiple jobs in parallel one of them will
fail and it's difficult to determine which one. This adds a new
V=e option to track when a job starts and stops. You can invoke
via:

 $ make -j24 V=e

You can run something like this now to see when each task begins
and completes:

 $ make -j24 V=e toolchain/uClibc/compile | awk '{ print strftime("%H:%M:%S"), $0; }'

Change-Id: Id91d48fd4f6552e94223fb66f26746d02b34036d
Signed-off-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>
Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>