[uClibc] build in parallel
uClibc is built during the toolchain step, there is nothing that can be
built at the same time until this is complete. It's also taking a while
to build because it's not using make -jX. So let's pass that through to
the build

After:
real	0m23.797s

Before:
real	1m32.835s

This should make the builds ~1 minute faster overall.

Change-Id: Ie72bc4a9419635cdd51d53377be22c57b754040a
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>