[ipq806x] Enable SPS and MMC_MSM_SPS_SUPPORT
SPS (Smart Peripheral Switch) is a DMA engine. It can move data in the following modes:
  1. Peripheral-to-Peripheral.
  2. Peripheral-to-Memory.
  3. Memory-to-Memory.

This is required for msm_sdcc driver. This is to scale up the frequency for the SDC1/SDC3
which is in dayotana fabric. By enabling this , the clock frequency will be scale up
dynamically switch to higher frequency to get the higher throughput.

CR-Fixed:CR546343

Change-Id: Ieabf36f0859b74914e33abd7a6b8d77599e7e384
Signed-off-by: Sham Muthayyan <smuthayy@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>