Create init script to manage power setting
The power settings will now be configurable through UCI using the
following key: system.powerctl.mode

For now, only the "auto" mode is supported, however this can be easily
extended to additional modes through the addition of new shell
functions. If the UCI key is missing, the script will default to "auto".

Change-Id: Ib5c181435b304063827f1a253e814a1224603f81
Signed-off-by: Pamidipati, Vijay <vpamidip@codeaurora.org>
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>