[feeds] Replace printing feeds by external printing stack
The default cups package in openwrt is v1.4.4. For cases like
USB connected printer, the libusb backend support is quite limited
in this version.

Morever, the install itself is broken, with basic commands such
as 'lpinfo' and 'lpstat' not returning any values.

With the openwrt-printing-packages, we not only get the new version
of cups, but we can also setup the entire printing stack
(including various drivers such as ghostscript, gutenprint etc.)
This will increase the image size over 100MB.  To keep the image small,
only the cups package is selected by default.

The idea is that the board will run cups with only the raw printer
driver, and will not modify the print stream in any way. The actual
printer driver must be installed at the source. The purpose of running
cups this way, is only to advertise the USB connected printer
over the network.

Change-Id: I5edb847fb895ff968fcb8c222a7f957eae3f8b19
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>
Signed-off-by: Srinivas Dodda <doddas@codeaurora.org>
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>