[qca-nss-drv] Move nss profiler option in the network dev menu
NSS Profiler do not define a lot of the options required in a package.
This change adds the same options as all the other NSS packages which
means it'll be in the same location.

It also reflects its dependency on the NSS driver by adding is to the
DEPENDS list.

Change-Id: I5567db303418c06bcbc5027325a40a73f7d051ec
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>