[openssl] Enable threading support at build time
When openssl is built with the no-threads option and used in
multi-threaded applications, asserts related to multi-threading are
reported.  The impact on the application is dependent of its
implementation and may cause different kinds of issues.

Here is an example of errors we get in this case:
md_rand.c: 325: ssleay_rand_add: Assertion `md_c[1] == md_count[1]' failed

Change-Id: I23ade9f0e0ff1bc610c0068db22a71fcc81188fa
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>