[ipq806x] Enable IRQ time accounting in config-3.4
In tickless kernels (CONFIG_NO_HZ), the values in /proc/stat do
not match the actual the actual load conditions. The reason being
that these values are dependent on the scheduler tick.

So the idea is to use the IRQ_TIME_ACCOUNTING feature
in the kernel. With this the accounting is partly done in interrupt
context.

Change-Id: I0533d5fb8a02e14e7a8f85ffc6bcf454a1c1ae71
CRs-Fixed: 553794
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>