[samba] Force user and group to root
By default, USB devices are mounted as root on OpenWrt. However, samba
uses the ID of the connected user to write files. So in case of
anonymous login, the user cannot write any file to the device.

The options "force user" and "force group" will not be set as "root" for
every share that we add into the system. As OpenWrt is a pretty simple
embedded system, it's using only one user "root" to do all its operation
therefore it doesn't compromise the system security as long as only the
mounted device is shared and that the rest of the FS is not visible from
the network.

CRs-Fixed: 537659
Change-Id: I2423e61d38d152697e0fd918f35145a94b6f3b5e
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>