[ipq806x] Disable ANDROID_PARANOID_NETWORK
ANDROID_PARANOID_NETWORK is used in Android to require specific uid/gid
to instanciate network sockets. As Android requires the App to obtain
certain priviledges before doing socket operations, it is a security
addition.

OpenWrt, on the other side, runs most daemons as root/root. Besides,
it's a router. So doing networking is really the point of every single
application in the system, and we should assume that instanciate
sockets will be required by all the daemons.

OpenWrt packages make the assumption that they can create sockets
without asking for specific priviledges. Some of them (like RIPd)
run as non-root user for security reasons; having this option enabled
is a problem in that case as they are not be able to create sockets,
therefore they fail.

So this change disables this option in our builds.

CRs-Fixed: 576065
Change-Id: Iec8e56f1b6d6689a4dd28fb52bba7223a6ee1c5f
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>