[ipq806x] Create /etc/sysctl.d before setting min_free_kbytes
The min_free_kbytes init doesn't test whether the /etc/sysctl.d
directory exists before creating it. If some package created it
before we're good but if nobody created it before the file will
not be created, therefore not setting the value in /proc.

This will create the directory with -p in case it doesn't exist.

Change-Id: I984ee510cafef361a8d5e7ff6f9cf978eec978a6
CRs-Fixed: 584010
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>