[ipq806x] Enable full preemption in the kernel
Full preemption adds better response time to the system and is a great
improvement for latency. In some cases, we want to have RT application
running on the device, therefore preemption will help getting to these
low-latency requirements.

The impact on performances has been measured to be < 1% whereas the
latency observed with this change is considerably lower (~20 times)
therefore we will make this change part of the default one.

Change-Id: Ida2cf544850cbc3bdd257d91edc7a576a66309f0
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>