[Config.in] Add an option to enable/disable DEBUG_INFO_REDUCED
CONFIG_DEBUG_INFO_REDUCED is set to "y" by default when enabling
CONFIG_DEBUG_INFO, which is not ideal as it reduces the amount of
debugging considerably. Therefore we add this option so we can
set/reset this option from OpenWrt and enable additional debugging
capabilities.

This is done in order to enhance the current debugging capabilities.

CRs-Fixed: 588735
Change-Id: If331e80d418bf13f4b7eabe77ef31861b5632ed8
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>