[ar71xx] ag71xx: allow force link up when phy_mask is null
The MDIO bus scan happens only if mii_bus->phy_mask is non-null
(drivers/net/phy/mdio_bus.c). If it is null, the MDIO generic driver
will assume that there is no device and therefore will not start the
scan.

In the ag71xx driver, mii_bus->phy_mask is directly copied from
ag71xx_platform_data->phy_mask. It is used by the MDIO driver to
start the scan and by the ag71xx driver to know on which PHY in the
switch we want to call phy_connect().

In cases where we have an external switch connected to the MDIO
interface and to only 1 xMII, we need the mdio bus to start a scan and
the GMAC link to be always up, which is not possible with the current
code.

We fix this problem by adding a new parameter in ag71xx_platform_data
(force_link) that can be populated in the platform file (mach-*) and
that can be used to avoid the ag71xx driver to call phy_connect(). The
MDIO probe will happen (phy_mask will be set to whatever PHY address we
can find the switch) and the driver will always use a fixed link.

Change-Id: I269ea8a421acfd4963c76713b88d096f6a80dc8e
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>