[base-files] Move VM config in an init.d file
VM config was done in rc.local, which is supposed to be an empty file
available for the user to customize its own boot process.

We'll move this logic in an init.d script. The main reason is we want
this configuration to be available sooner in the boot process (rc.local
is executed at the end) and we want the rc.local to be used the way
it was designed, i.e. by the user.

Change-Id: I34e4e564b4e52a5d0d5e74ba5bd8602fb422211a
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>