[ipq806x] Make VM overcommit policy specific to ipq806x
Stability and stress tests show unstabilities in extreme test cases when
using overcommit on IPQ806x. Therefore we disabled overcommit and set
the overcommit_ratio to 50%.

This value was empirical, but it creates issues on low-memory platforms
(<=128M) as it doesn't leave enough room to the applications to run
correctly. Therefore we will make it platform specific by moving it to
the ipq806x folder.

For now, it's an acceptable trade-off, but over time we need to figure
out what's the right system-wide policy. There is no technical reason
which could rightfully explain why overcommit is good on some platforms
and not on others. It is a system-wide setting and should be set by the
distribution, not by the platform.

Change-Id: Ibf06f2a8a8bb4acc98336b26bae068c16dec697d
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>