[ipq806x] Include kernel uImage in UBI image
This change is a preliminary step to move the kernel image into the UBI
layer so it gets all the bad block management and wear levelling and be
more robust.

This doesn't impact the flash layout for now, but will be required when
we will make the layout change.

CRs-Fixed: 623078
Change-Id: I944a85d8cdf6e1133bdd41a46484c3fbdc969432
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>