[scripts] metadata.pl: fix dependencies report when using PROVIDES
When a package has a conditional dependency on a package which is using
the PROVIDES mechanism, the condition is not reported in the Config.in.
However, the dependencies are reported, which results in packages
depending on other packages directly whereas they should really depend
on them only if another package has been selected.

Change-Id: I0f96ff852ec66fd248d843e3597a1fdc109b4230
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>