[ipq806x] Include kernel uImage in UBI image
This change is a preliminary step to move the kernel image into the UBI
layer so it gets all the bad block management and wear levelling and be
more robust.

This doesn't impact the flash layout for now, but will be required when
we will make the layout change.

Change-Id: I2458679c3de945d51ad91ce146d753a5ca6a03d5
CRs-Fixed: 623078
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Signed-off-by: Mohammmed Shafi Shajakhan <mohammed@codeaurora.org>