include: Clean-up kernel git command options management
The way we're currently managing the git options in kernel-defaults.mk
is not scalable. If requires different code path for each option; it's
ok so far as we handle only one option, but if we want to make the git
clone mechanism more flexible, we'll need to add more option, which
will become tedious.

So; we'll move the GIT options into a variable, that may or may not be
set depending on the configuration, and we'll pass this variable to the
git command.

Change-Id: I4a93db65f7270156cc4eb73dbd1c826942a70916
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>