include: Allow OpenWrt patch application on git kernel
By default, OpenWrt doesn't patch the kernel if it's downloaded from
a git tree. This option will allow OpenWrt to download the kernel from
GIT but still allow the user to apply some patches, when enabled.

It can be very convenient when working on some standard kernel
maintained in Git, shared by multiple Linux distributions, but we still
want to take advantages of all the OpenWrt patches.

Change-Id: I229444e0052977fe3260484d1ac74627d0946e7f
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>