[net] Reduce NET_SKB_PAD to L1_CACHE_BYTES
NET_SKB_PAD was originally increased to 64B to cacheline-align the
request.  At the time, it could not be decreased to align it, as the
wifi 802.11n driver would kernel panic.
Recent testing has revealed that this increase is no longer needed.
This change reduces NET_SKB_PAD to the next lower L1_CACHE_BYTES-
aligned value.

Change-Id: I31a049f502043d61da9262296b48085e154f01c9
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>