ipq806x: Use "Image" file rather than "zImage"
zImage file contains the relocator, which we don't want to use in our
case due to the NSS memory constraint. Instead, we want to use the
Image and set the load address statically at packaging time through
the mkimage variable. It has to match what is defined in the kernel.

Change-Id: I008b1e1fe1e55d38483d711efe7a2806ec0a1fab
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>