[package-defaults] remove file before overwritting it
config.{sub,guess} could be symlinks to a shared common version of
this file (e.g. in staging). so we remove the destination file via
--remove-destination option of cp. this prevents replaceing the
common file that other packages could be build with if running at
the same time

this fixes a class of errors where config.sub is missing, or
only partially present when running configure

Change-Id: Iad4b15cf016ce8d296787355fe0247ad488d9b4e
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>