build: remove file before overwritting it with replace_script
config.{sub,guess} could be symlinks to a shared common version of
this file (e.g. in staging). So we remove the destination file via
--remove-destination option of cp. This prevents replaceing the
common file that other packages could be build with if running at
the same time.

This fixes a class of errors where config.sub is missing, or
only partially present when running configure because a cp is
currently in progress

This is commonly seen building with a lot of parallel jobs and
on packages that use 'PKG_FIXUP:=autoreconf'

Change-Id: Iad4b15cf016ce8d296787355fe0247ad488d9b4e
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>