ipq806x: Fix modules AUTOLOAD & clean-up kernel config
AUTOLOAD was broken due to the presence of the kernel option
CONFIG_LOCALVERSION_AUTO=y in the kernel config. As we use a git clone
command to generate the kernel tree, that causes the version string
to include the sha1, therefore it didn't match the directory name
in /lib/modules/... when kmodloader tried to detect the modules.

This change reverted a lot of kernel options to their OpenWrt default,
getting the kernel configuration file to be closer to what OpenWrt would
expect. This is done to prevent future issues.

Change-Id: I7e6a5fd8b32baa0073ec97a02fbb753205c58da8
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>