[coccinelle] Add new coccinelle package
Coccinelle is used, in particular to generate the backport. As using
the bleeding edge backports package requires to use the bleeding edge
coccinelle package, we'll facilitate the deployment by building it
from OpenWrt rather than requiring each individual to forward-port
the package for its distribution.

This change requires a new tool to be installed on the machine however,
ocaml-nox which is present in most distributions.

Change-Id: I61ca4a2c6db16b6ae5675f05025a9408e2f6369c
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>