hostapd: Add optional support for hostapd own_ip_addr
in wireless config
`own_ip_addr` is used by hostapd as NAS-IP-Address.
This is used to identify the AP that is requesting
the authentication of the user and could be used to
define which AP's can authenticate users.
Some vendors implement only NAS-Identifier or NAS-IP-Address
and not both.
This patch adds ownip as an optional parameter in
/etc/config/wireless.

Signed-off-by: Thomas Wouters <thomaswouters@gmail.com>

Change-Id: Icb40a8e12d3e7e675d705761eb8dde279b699ee0
git-svn-id: svn://svn.openwrt.org/openwrt/trunk@40934 3c298f89-4303-0410-b956-a3cf2f4a3e73
Signed-off-by: Peter Oh <poh@codeaurora.org>